# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI & ML Learning Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export type TrackDifficulty =
  (typeof TrackDifficulty)[keyof typeof TrackDifficulty];

export const TrackDifficulty = {
  beginner: "beginner",
  intermediate: "intermediate",
  advanced: "advanced",
  expert: "expert",
} as const;

export interface ModuleSummary {
  id: string;
  title: string;
  description: string;
  lessonCount: number;
  estimatedHours: number;
  order: number;
}

export interface Track {
  id: string;
  title: string;
  description: string;
  icon?: string;
  difficulty: TrackDifficulty;
  estimatedHours: number;
  moduleCount: number;
  lessonCount: number;
  tags: string[];
  modules: ModuleSummary[];
  color?: string;
  order: number;
}

export type LessonSummaryType =
  (typeof LessonSummaryType)[keyof typeof LessonSummaryType];

export const LessonSummaryType = {
  concept: "concept",
  coding: "coding",
  quiz: "quiz",
  project: "project",
} as const;

export interface LessonSummary {
  id: string;
  title: string;
  description: string;
  estimatedMinutes: number;
  type: LessonSummaryType;
  order: number;
}

export interface Module {
  id: string;
  trackId: string;
  title: string;
  description: string;
  order: number;
  estimatedHours: number;
  lessons: LessonSummary[];
}

export type LessonType = (typeof LessonType)[keyof typeof LessonType];

export const LessonType = {
  concept: "concept",
  coding: "coding",
  quiz: "quiz",
  project: "project",
} as const;

export type LessonSectionType =
  (typeof LessonSectionType)[keyof typeof LessonSectionType];

export const LessonSectionType = {
  text: "text",
  code: "code",
  math: "math",
  visualization: "visualization",
  callout: "callout",
  quiz_inline: "quiz_inline",
} as const;

export type LessonSectionCalloutType =
  (typeof LessonSectionCalloutType)[keyof typeof LessonSectionCalloutType];

export const LessonSectionCalloutType = {
  info: "info",
  warning: "warning",
  tip: "tip",
  important: "important",
} as const;

export interface LessonSection {
  id: string;
  title?: string;
  type: LessonSectionType;
  content: string;
  language?: string;
  caption?: string;
  calloutType?: LessonSectionCalloutType;
}

export type ExerciseType = (typeof ExerciseType)[keyof typeof ExerciseType];

export const ExerciseType = {
  multiple_choice: "multiple_choice",
  fill_code: "fill_code",
  write_code: "write_code",
  true_false: "true_false",
} as const;

export interface Exercise {
  id: string;
  type: ExerciseType;
  question: string;
  description?: string;
  starterCode?: string;
  solutionCode?: string;
  options?: string[];
  correctAnswer?: string;
  explanation?: string;
  hints?: string[];
}

export interface Lesson {
  id: string;
  moduleId: string;
  trackId: string;
  title: string;
  description: string;
  type: LessonType;
  estimatedMinutes: number;
  order: number;
  sections: LessonSection[];
  exercises: Exercise[];
  keyTakeaways: string[];
  prerequisites: string[];
  nextLessonId?: string;
  prevLessonId?: string;
}

export type LessonProgressStatus =
  (typeof LessonProgressStatus)[keyof typeof LessonProgressStatus];

export const LessonProgressStatus = {
  not_started: "not_started",
  in_progress: "in_progress",
  completed: "completed",
} as const;

export interface LessonProgress {
  lessonId: string;
  status: LessonProgressStatus;
  completedAt?: string;
  score?: number;
  timeSpentMinutes?: number;
}

export interface ExerciseResult {
  exerciseId: string;
  isCorrect: boolean;
  answer: string;
  explanation: string;
  xpEarned: number;
  submittedAt?: string;
}

export interface UserProgress {
  completedLessons: string[];
  lessonProgress: LessonProgress[];
  exerciseResults: ExerciseResult[];
  totalXp: number;
  streakDays: number;
  lastActiveDate?: string;
}

export type LessonProgressUpdateStatus =
  (typeof LessonProgressUpdateStatus)[keyof typeof LessonProgressUpdateStatus];

export const LessonProgressUpdateStatus = {
  not_started: "not_started",
  in_progress: "in_progress",
  completed: "completed",
} as const;

export interface LessonProgressUpdate {
  status: LessonProgressUpdateStatus;
  timeSpentMinutes?: number;
}

export interface ExerciseSubmission {
  answer: string;
  code?: string;
  timeSpentSeconds?: number;
}