# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AI & ML Learning Platform API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: curriculum
    description: Curriculum and course operations
  - name: progress
    description: User progress tracking
  - name: lessons
    description: Lesson content operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /curriculum/tracks:
    get:
      operationId: getTracks
      tags: [curriculum]
      summary: Get all learning tracks
      description: Returns all available learning tracks with their modules
      responses:
        "200":
          description: List of tracks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Track"

  /curriculum/tracks/{trackId}:
    get:
      operationId: getTrack
      tags: [curriculum]
      summary: Get a specific track
      parameters:
        - name: trackId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Track details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Track"
        "404":
          description: Track not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /curriculum/modules/{moduleId}:
    get:
      operationId: getModule
      tags: [curriculum]
      summary: Get a specific module with its lessons
      parameters:
        - name: moduleId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Module with lessons
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Module"
        "404":
          description: Module not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /curriculum/lessons/{lessonId}:
    get:
      operationId: getLesson
      tags: [curriculum]
      summary: Get a full lesson with content, examples and exercises
      parameters:
        - name: lessonId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Full lesson content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lesson"
        "404":
          description: Lesson not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /progress:
    get:
      operationId: getProgress
      tags: [progress]
      summary: Get all user progress
      responses:
        "200":
          description: User progress data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProgress"

  /progress/lesson/{lessonId}:
    post:
      operationId: updateLessonProgress
      tags: [progress]
      summary: Update progress for a specific lesson
      parameters:
        - name: lessonId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LessonProgressUpdate"
      responses:
        "200":
          description: Updated progress
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LessonProgress"

  /progress/exercise/{exerciseId}:
    post:
      operationId: submitExercise
      tags: [progress]
      summary: Submit an exercise answer
      parameters:
        - name: exerciseId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExerciseSubmission"
      responses:
        "200":
          description: Exercise result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExerciseResult"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message

    Track:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        icon:
          type: string
        difficulty:
          type: string
          enum: [beginner, intermediate, advanced, expert]
        estimatedHours:
          type: number
        moduleCount:
          type: number
        lessonCount:
          type: number
        tags:
          type: array
          items:
            type: string
        modules:
          type: array
          items:
            $ref: "#/components/schemas/ModuleSummary"
        color:
          type: string
        order:
          type: integer
      required:
        - id
        - title
        - description
        - difficulty
        - estimatedHours
        - moduleCount
        - lessonCount
        - tags
        - modules
        - order

    ModuleSummary:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        lessonCount:
          type: number
        estimatedHours:
          type: number
        order:
          type: integer
      required:
        - id
        - title
        - description
        - lessonCount
        - estimatedHours
        - order

    Module:
      type: object
      properties:
        id:
          type: string
        trackId:
          type: string
        title:
          type: string
        description:
          type: string
        order:
          type: integer
        estimatedHours:
          type: number
        lessons:
          type: array
          items:
            $ref: "#/components/schemas/LessonSummary"
      required:
        - id
        - trackId
        - title
        - description
        - order
        - estimatedHours
        - lessons

    LessonSummary:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        estimatedMinutes:
          type: number
        type:
          type: string
          enum: [concept, coding, quiz, project]
        order:
          type: integer
      required:
        - id
        - title
        - description
        - estimatedMinutes
        - type
        - order

    Lesson:
      type: object
      properties:
        id:
          type: string
        moduleId:
          type: string
        trackId:
          type: string
        title:
          type: string
        description:
          type: string
        type:
          type: string
          enum: [concept, coding, quiz, project]
        estimatedMinutes:
          type: number
        order:
          type: integer
        sections:
          type: array
          items:
            $ref: "#/components/schemas/LessonSection"
        exercises:
          type: array
          items:
            $ref: "#/components/schemas/Exercise"
        keyTakeaways:
          type: array
          items:
            type: string
        prerequisites:
          type: array
          items:
            type: string
        nextLessonId:
          type: string
        prevLessonId:
          type: string
      required:
        - id
        - moduleId
        - trackId
        - title
        - description
        - type
        - estimatedMinutes
        - order
        - sections
        - exercises
        - keyTakeaways
        - prerequisites

    LessonSection:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        type:
          type: string
          enum: [text, code, math, visualization, callout, quiz_inline]
        content:
          type: string
        language:
          type: string
        caption:
          type: string
        calloutType:
          type: string
          enum: [info, warning, tip, important]
      required:
        - id
        - type
        - content

    Exercise:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
          enum: [multiple_choice, fill_code, write_code, true_false]
        question:
          type: string
        description:
          type: string
        starterCode:
          type: string
        solutionCode:
          type: string
        options:
          type: array
          items:
            type: string
        correctAnswer:
          type: string
        explanation:
          type: string
        hints:
          type: array
          items:
            type: string
      required:
        - id
        - type
        - question

    UserProgress:
      type: object
      properties:
        completedLessons:
          type: array
          items:
            type: string
        lessonProgress:
          type: array
          items:
            $ref: "#/components/schemas/LessonProgress"
        exerciseResults:
          type: array
          items:
            $ref: "#/components/schemas/ExerciseResult"
        totalXp:
          type: number
        streakDays:
          type: number
        lastActiveDate:
          type: string
      required:
        - completedLessons
        - lessonProgress
        - exerciseResults
        - totalXp
        - streakDays

    LessonProgress:
      type: object
      properties:
        lessonId:
          type: string
        status:
          type: string
          enum: [not_started, in_progress, completed]
        completedAt:
          type: string
        score:
          type: number
        timeSpentMinutes:
          type: number
      required:
        - lessonId
        - status

    LessonProgressUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [not_started, in_progress, completed]
        timeSpentMinutes:
          type: number
      required:
        - status

    ExerciseSubmission:
      type: object
      properties:
        answer:
          type: string
        code:
          type: string
        timeSpentSeconds:
          type: number
      required:
        - answer

    ExerciseResult:
      type: object
      properties:
        exerciseId:
          type: string
        isCorrect:
          type: boolean
        answer:
          type: string
        explanation:
          type: string
        xpEarned:
          type: number
        submittedAt:
          type: string
      required:
        - exerciseId
        - isCorrect
        - answer
        - explanation
        - xpEarned