# lib/api-zod/src/generated/types/exercise.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI & ML Learning Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { ExerciseType } from "./exerciseType";

export interface Exercise {
  id: string;
  type: ExerciseType;
  question: string;
  description?: string;
  starterCode?: string;
  solutionCode?: string;
  options?: string[];
  correctAnswer?: string;
  explanation?: string;
  hints?: string[];
}