# lib/api-zod/src/generated/types/lesson.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI & ML Learning Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { Exercise } from "./exercise";
import type { LessonSection } from "./lessonSection";
import type { LessonType } from "./lessonType";

export interface Lesson {
  id: string;
  moduleId: string;
  trackId: string;
  title: string;
  description: string;
  type: LessonType;
  estimatedMinutes: number;
  order: number;
  sections: LessonSection[];
  exercises: Exercise[];
  keyTakeaways: string[];
  prerequisites: string[];
  nextLessonId?: string;
  prevLessonId?: string;
}