# lib/api-zod/src/generated/types/lessonProgress.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI & ML Learning Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { LessonProgressStatus } from "./lessonProgressStatus";

export interface LessonProgress {
  lessonId: string;
  status: LessonProgressStatus;
  completedAt?: string;
  score?: number;
  timeSpentMinutes?: number;
}