# lib/api-zod/src/generated/types/lessonProgressUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI & ML Learning Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { LessonProgressUpdateStatus } from "./lessonProgressUpdateStatus";

export interface LessonProgressUpdate {
  status: LessonProgressUpdateStatus;
  timeSpentMinutes?: number;
}