# lib/api-zod/src/generated/types/lessonSection.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI & ML Learning Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { LessonSectionCalloutType } from "./lessonSectionCalloutType";
import type { LessonSectionType } from "./lessonSectionType";

export interface LessonSection {
  id: string;
  title?: string;
  type: LessonSectionType;
  content: string;
  language?: string;
  caption?: string;
  calloutType?: LessonSectionCalloutType;
}