# lib/api-zod/src/generated/types/module.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI & ML Learning Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { LessonSummary } from "./lessonSummary";

export interface Module {
  id: string;
  trackId: string;
  title: string;
  description: string;
  order: number;
  estimatedHours: number;
  lessons: LessonSummary[];
}