# lib/api-zod/src/generated/types/track.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI & ML Learning Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { ModuleSummary } from "./moduleSummary";
import type { TrackDifficulty } from "./trackDifficulty";

export interface Track {
  id: string;
  title: string;
  description: string;
  icon?: string;
  difficulty: TrackDifficulty;
  estimatedHours: number;
  moduleCount: number;
  lessonCount: number;
  tags: string[];
  modules: ModuleSummary[];
  color?: string;
  order: number;
}