# lib/api-zod/src/generated/types/userProgress.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI & ML Learning Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { ExerciseResult } from "./exerciseResult";
import type { LessonProgress } from "./lessonProgress";

export interface UserProgress {
  completedLessons: string[];
  lessonProgress: LessonProgress[];
  exerciseResults: ExerciseResult[];
  totalXp: number;
  streakDays: number;
  lastActiveDate?: string;
}